Fix test skip if python binaries missing
On Debian based system, the Python binary includes the version and is
named `python3`. The test cases that expect that a binary named `python`
exists will fail.

Use the binary name of the running Python process for the test cases to
ensure that this binary is present.

Forwarded: https://github.com/saltstack/pytest-salt-factories/pull/53
Signed-off-by: Benjamin Drung <benjamin.drung@ionos.com>